pyproject.toml
[project]
name = "harborline"
version = "0.1.0"
description = "Unified commerce core for orders, payments, inventory, and real-time ops."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Harborline Team" }
]
keywords = ["fastapi", "commerce", "graphql", "websocket", "hoppscotch"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.111,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "PyJWT>=2.8,<3.0",
    "python-multipart>=0.0.9,<1.0",
    "strawberry-graphql>=0.247,<1.0",
    "jinja2>=3.1,<4.0",
    "pydantic-settings>=2.2,<3.0",
    "sqlalchemy>=2.0,<3.0",
    "psycopg[binary]>=3.1,<4.0",
    "slowapi>=0.1.9,<1.0",
    "opentelemetry-api>=1.25,<2.0",
    "opentelemetry-sdk>=1.25,<2.0",
    "opentelemetry-exporter-otlp-proto-http>=1.25,<2.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0,<1.0",
    "opentelemetry-instrumentation-sqlalchemy>=0.46b0,<1.0",
    "neo4j>=5.10,<6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-cov>=4.0,<6.0",
    "pytest-asyncio>=0.23,<1.0",
    "httpx>=0.27,<1.0",
    "ruff>=0.4,<1.0",
    "mypy>=1.10,<2.0",
    "pre-commit>=3.7,<4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:strawberry.*",
    "ignore:.*HMAC key.*:UserWarning",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
